Add MATH 525 Notes
Originally, I wanted to add a notes tab on the page and
list out all my notes there but I wasn't able to get the context
of it with Hakyll and it ended up giving me a bunch of headaches. So,
for now I'll just store everything on the homepage!